+++ contact_modification/__manifest__.py
@@ -8,7 +8,7 @@
     'summary': 'To manage Contacts with classification',
     'author': 'CLx Media',
     'website': 'https://conversionlogix.com/',
-    'version': '13.1.1.0.26',
+    'version': '13.1.1.0.27',
     'sequence': 1,
     'license': 'AGPL-3',
     'depends': [